+++ spec/praxis/mapper/resource_spec.rb
@@ -16,15 +16,19 @@ describe Praxis::Mapper::Resource do
       subject(:properties) { resource.properties }
 
       it 'includes directly-set properties' do
-        expect(properties[:other_resource]).to eq(dependencies: [:other_model], through: nil)
+        expect(properties[:other_resource]).to eq(dependencies: [:other_model], through: nil, as: :other_resource)
+      end
+
+      it 'includes aliases as well if different from name' do
+        expect(properties[:aliased_association]).to eq(dependencies: [:name], through: nil, as: :other_model)
       end
 
       it 'inherits from a superclass' do
-        expect(properties[:href]).to eq(dependencies: [:id], through: nil)
+        expect(properties[:href]).to eq(dependencies: [:id], through: nil, as: :href)
       end
 
       it 'properly overrides a property from the parent' do
-        expect(properties[:name]).to eq(dependencies: [:simple_name], through: nil)
+        expect(properties[:name]).to eq(dependencies: [:simple_name], through: nil, as: :name)
       end
     end
   end